Quick and safe setup steps.
after getting new usb hub, thought it would be fun to play around and after recently knowing about SOLID principles, thought to edit the code in the REcordProvider but seems like there is no need. ANd also, Thought to make this code available offline too, but you can't fetch the user histories and all without the user being logged in, and the user can't be logged in unless you are online. Plus, I could do a display message for you are not online, but I am not doing such noob stuff anymore.